+++ lib/pages/habits/habits_page.dart
@@ -8,7 +8,7 @@ import 'package:lotti/pages/dashboards/dashboard_page.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/platform.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
-import 'package:lotti/widgets/charts/dashboard_habits_chart.dart';
+import 'package:lotti/widgets/charts/habits/dashboard_habits_chart.dart';
 import 'package:lotti/widgets/charts/habits/habit_completion_rate_chart.dart';
 import 'package:lotti/widgets/charts/utils.dart';