+++ lib/pages/habits/habits_page.dart
@@ -8,7 +8,8 @@ import 'package:lotti/pages/dashboards/dashboard_page.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/platform.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
-import 'package:lotti/widgets/charts/dashboard_habits_chart.dart';
+import 'package:lotti/widgets/charts/habits/dashboard_habits_chart.dart';
+import 'package:lotti/widgets/charts/habits/habit_completion_rate_chart.dart';
 import 'package:lotti/widgets/charts/utils.dart';
 
 class HabitsTabPage extends StatelessWidget {
@@ -54,11 +55,11 @@ class HabitsTabPage extends StatelessWidget {
                       onValueChanged: cubit.setTimeSpan,
                       children: {
                         if (isMobile) 7: const DaysSegment('7'),
-                        14: const DaysSegment('14'),
-                        30: const DaysSegment('30'),
-                        90: const DaysSegment('90'),
+                        14: const DaysSegment('14 d'),
+                        30: const DaysSegment('30 d'),
+                        90: const DaysSegment('90 d'),
                         if (isDesktop || landscape)
-                          180: const DaysSegment('180'),
+                          180: const DaysSegment('180 d'),
                       },
                     ),
                   ),
@@ -161,7 +162,7 @@ class HabitsPageAppBar extends StatelessWidget with PreferredSizeWidget {
   HabitsPageAppBar({super.key});
 
   @override
-  Size get preferredSize => const Size.fromHeight(kToolbarHeight + 40);
+  Size get preferredSize => const Size.fromHeight(kToolbarHeight + 110);
 
   @override
   Widget build(BuildContext context) {
@@ -179,83 +180,8 @@ class HabitsPageAppBar extends StatelessWidget with PreferredSizeWidget {
         return Column(
           children: [
             TitleAppBar(title: '$title $habitCounters', showBackButton: false),
-            const HabitsPageProgressBar(),
-          ],
-        );
-      },
-    );
-  }
-}
-
-class HabitsPageProgressBar extends StatelessWidget {
-  const HabitsPageProgressBar({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<HabitsCubit, HabitsState>(
-      builder: (context, HabitsState state) {
-        final width = MediaQuery.of(context).size.width - 200;
-        const height = 20.0;
-        final total = state.habitDefinitions.length;
-
-        if (total == 0) {
-          return const SizedBox.shrink();
-        }
-
-        final successfulToday = state.successfulToday.length;
-        final done = successfulToday / total;
-        final percentage = (done * 100).toInt();
-
-        final greenWidth = done * width;
-        final redWidth = width - greenWidth - 1;
-
-        return Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Opacity(
-              opacity: 0,
-              child: Padding(
-                padding: const EdgeInsets.only(left: 20),
-                child: Text(
-                  '$percentage %',
-                  style: chartTitleStyle().copyWith(
-                    color: styleConfig().primaryColor,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-            ),
-            ClipRRect(
-              borderRadius: BorderRadius.circular(height),
-              child: Row(
-                children: [
-                  Container(
-                    width: greenWidth,
-                    height: height,
-                    color: styleConfig().primaryColor,
-                  ),
-                  const SizedBox(width: 1),
-                  Opacity(
-                    opacity: 0.5,
-                    child: Container(
-                      width: redWidth,
-                      height: height,
-                      color: styleConfig().alarm,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(left: 20),
-              child: Text(
-                '$percentage %',
-                style: chartTitleStyle().copyWith(
-                  color: styleConfig().primaryColor,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
+            //const HabitsPageProgressBar(),
+            const HabitCompletionRateChart(),
           ],
         );
       },