+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 - Chart for habit completion rate over time
 
+### Changed:
+- Improved whitespace in habit success indicators
+
 ## [0.8.204] - 2022-12-07
 ### Added:
 - Progress bar for habit progress for the current day